Implementando ajustes e logica do requisito funcional do sistema e implementando testes unitarios

+++ src/main/java/com/mazza/tech/gestao/estacionamento/dto/GarageDTO.java
@@ -1,12 +1,13 @@
 package com.mazza.tech.gestao.estacionamento.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
 
 @Data
 public class GarageDTO {
- private String name;
- private double basePrice;
- private List<SpotDTO> spots;
+	private String name;
+	private BigDecimal basePrice;
+	private List<SpotDTO> spots;
 }
